refactor(ui-kit): continue shared component migration
- migrate repeated CTAs and empty states to shared primitives

- add UI-kit guards for remaining first-wave migrations
